# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentHub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Subject {
  id: number;
  name: string;
  /** @nullable */
  teacher?: string | null;
  color: string;
  /** @nullable */
  description?: string | null;
  archived: boolean;
  createdAt: string;
}

export interface SubjectInput {
  name: string;
  /** @nullable */
  teacher?: string | null;
  color: string;
  /** @nullable */
  description?: string | null;
  archived?: boolean;
}

export interface SubjectWithStats {
  id: number;
  name: string;
  /** @nullable */
  teacher?: string | null;
  color: string;
  /** @nullable */
  description?: string | null;
  archived: boolean;
  /** @nullable */
  averageGrade: number | null;
  debtCount: number;
  taskCount: number;
  upcomingTestCount: number;
  /** @nullable */
  nextLesson: string | null;
}

export interface Lesson {
  id: number;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  subjectColor?: string | null;
  title: string;
  startsAt: string;
  endsAt: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  teacher?: string | null;
  source: string;
  /** @nullable */
  description?: string | null;
}

export interface Grade {
  id: number;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  subjectColor?: string | null;
  value: number;
  maxValue: number;
  kind: string;
  receivedAt: string;
  weight: number;
  /** @nullable */
  comment?: string | null;
}

export interface Debt {
  id: number;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  subjectColor?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  appearedAt: string;
  /** @nullable */
  deadline?: string | null;
  status: string;
  importance: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  comment?: string | null;
}

export interface Test {
  id: number;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  subjectColor?: string | null;
  title: string;
  scheduledAt: string;
  /** @nullable */
  deadline?: string | null;
  status: string;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  link?: string | null;
  importance: string;
  needsPreparation: boolean;
  /** @nullable */
  preparationPlan?: string | null;
}

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  subjectColor?: string | null;
  status: string;
  priority: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  estimatedMinutes?: number | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  subjectColor?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface SubjectDetail {
  subject: Subject;
  /** @nullable */
  averageGrade: number | null;
  lessons: Lesson[];
  grades: Grade[];
  debts: Debt[];
  tests: Test[];
  tasks: Task[];
  notes: Note[];
}

export interface LessonInput {
  /** @nullable */
  subjectId?: number | null;
  title: string;
  startsAt: string;
  endsAt: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  teacher?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface TaskInput {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subjectId?: number | null;
  status?: string;
  priority?: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  estimatedMinutes?: number | null;
}

export interface DebtInput {
  /** @nullable */
  subjectId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  deadline?: string | null;
  status?: string;
  importance?: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  comment?: string | null;
}

export interface GradeInput {
  /** @nullable */
  subjectId?: number | null;
  value: number;
  maxValue: number;
  kind: string;
  receivedAt: string;
  weight?: number;
  /** @nullable */
  comment?: string | null;
}

export type GradeSummaryBySubjectItem = {
  subjectId: number;
  subjectName: string;
  subjectColor: string;
  /** @nullable */
  average: number | null;
  count: number;
};

export interface GradeSummary {
  /** @nullable */
  overallAverage: number | null;
  bySubject: GradeSummaryBySubjectItem[];
}

export interface TestInput {
  /** @nullable */
  subjectId?: number | null;
  title: string;
  scheduledAt: string;
  /** @nullable */
  deadline?: string | null;
  status?: string;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  link?: string | null;
  importance?: string;
  needsPreparation?: boolean;
  /** @nullable */
  preparationPlan?: string | null;
}

export interface NoteInput {
  title: string;
  content: string;
  /** @nullable */
  subjectId?: number | null;
  tags?: string[];
}

export interface WeeklyPlan {
  id: number;
  weekStartDate: string;
  goals: string;
  notes: string;
  status: string;
  updatedAt: string;
}

export interface WeeklyPlanInput {
  weekStartDate: string;
  goals: string;
  notes: string;
  status: string;
}

export interface FreeSlot {
  startsAt: string;
  endsAt: string;
  durationMinutes: number;
}

export interface DashboardStats {
  tasksToday: number;
  tasksDoneToday: number;
  tasksTotal: number;
  tasksDoneTotal: number;
  openDebts: number;
  upcomingTests7d: number;
  lessonsToday: number;
  /** @nullable */
  averageGrade: number | null;
}

export interface DashboardToday {
  date: string;
  lessons: Lesson[];
  tasks: Task[];
  debts: Debt[];
  upcomingTests: Test[];
  freeSlots: FreeSlot[];
  stats: DashboardStats;
}

export interface Deadline {
  id: string;
  kind: string;
  title: string;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  subjectColor?: string | null;
  dueAt: string;
  importance: string;
}

export interface WeekDay {
  date: string;
  weekday: number;
  lessons: Lesson[];
  tasks: Task[];
  tests: Test[];
  debts: Debt[];
  loadScore: number;
}

export interface WeekView {
  weekStart: string;
  weekEnd: string;
  days: WeekDay[];
}

export type GetFreeSlotsParams = {
  date?: string;
};

export type GetWeekParams = {
  weekStart?: string;
};

export type ListLessonsParams = {
  subjectId?: number;
  from?: string;
  to?: string;
};

export type ListTasksParams = {
  subjectId?: number;
  status?: string;
  dueDate?: string;
};

export type ListDebtsParams = {
  subjectId?: number;
  status?: string;
};

export type ListGradesParams = {
  subjectId?: number;
};

export type ListTestsParams = {
  subjectId?: number;
  status?: string;
};

export type ListNotesParams = {
  subjectId?: number;
};