# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StudentHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/free-slots:
    get:
      operationId: getFreeSlots
      tags: [dashboard]
      summary: Free time slots in the day
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          description: Free slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FreeSlot"
  /week:
    get:
      operationId: getWeek
      tags: [dashboard]
      summary: Week view with lessons, tasks, tests grouped by day
      parameters:
        - name: weekStart
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          description: Week view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekView"
...
  /debts:
    get:
      operationId: listDebts
      tags: [debts]
      parameters:
        - { name: subjectId, in: query, required: false, schema: { type: integer } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Debts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Debt" }
    post:
      operationId: createDebt
      tags: [debts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DebtInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Debt" }
  /debts/{id}:
    patch:
      operationId: updateDebt
      tags: [debts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DebtInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Debt" }
    delete:
      operationId: deleteDebt
      tags: [debts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /grades:
    get:
      operationId: listGrades
      tags: [grades]
      parameters:
        - { name: subjectId, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: Grades
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Grade" }
    post:
      operationId: createGrade
      tags: [grades]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GradeInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Grade" }
  /grades/{id}:
    delete:
      operationId: deleteGrade
      tags: [grades]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Lesson:
      type: object
      properties:
        id: { type: integer }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        subjectColor: { type: ["string", "null"] }
        title: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        location: { type: ["string", "null"] }
        teacher: { type: ["string", "null"] }
        source: { type: string }
        description: { type: ["string", "null"] }
      required: [id, title, startsAt, endsAt, source]
...
    Task:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        subjectColor: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        dueDate: { type: ["string", "null"], format: date-time }
        estimatedMinutes: { type: ["integer", "null"] }
        completedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, title, status, priority, createdAt]
...
    Debt:
      type: object
      properties:
        id: { type: integer }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        subjectColor: { type: ["string", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        appearedAt: { type: string, format: date-time }
        deadline: { type: ["string", "null"], format: date-time }
        status: { type: string }
        importance: { type: string }
        link: { type: ["string", "null"] }
        comment: { type: ["string", "null"] }
      required: [id, title, status, importance, appearedAt]
    DebtInput:
      type: object
      properties:
        subjectId: { type: ["integer", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        deadline: { type: ["string", "null"], format: date-time }
        status: { type: string }
        importance: { type: string }
        link: { type: ["string", "null"] }
        comment: { type: ["string", "null"] }
      required: [title]
    Grade:
      type: object
      properties:
        id: { type: integer }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        subjectColor: { type: ["string", "null"] }
        value: { type: number }
        maxValue: { type: number }
        kind: { type: string }
        receivedAt: { type: string, format: date }
        weight: { type: number }
        comment: { type: ["string", "null"] }
      required: [id, value, maxValue, kind, receivedAt, weight]
    GradeInput:
      type: object
      properties:
        subjectId: { type: ["integer", "null"] }
        value: { type: number }
        maxValue: { type: number }
        kind: { type: string }
        receivedAt: { type: string, format: date }
        weight: { type: number }
        comment: { type: ["string", "null"] }
      required: [value, maxValue, kind, receivedAt]
...
    Test:
      type: object
      properties:
        id: { type: integer }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        subjectColor: { type: ["string", "null"] }
        title: { type: string }
        scheduledAt: { type: string, format: date-time }
        deadline: { type: ["string", "null"], format: date-time }
        status: { type: string }
        result: { type: ["string", "null"] }
        link: { type: ["string", "null"] }
        importance: { type: string }
        needsPreparation: { type: boolean }
        preparationPlan: { type: ["string", "null"] }
      required: [id, title, scheduledAt, status, importance, needsPreparation]
...
    FreeSlot:
      type: object
      properties:
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        durationMinutes: { type: integer }
      required: [startsAt, endsAt, durationMinutes]
    WeekView:
      type: object
      properties:
        weekStart: { type: string, format: date }
        weekEnd: { type: string, format: date }
        days:
          type: array
          items: { $ref: "#/components/schemas/WeekDay" }
      required: [weekStart, weekEnd, days]
    WeekDay:
      type: object
      properties:
        date: { type: string, format: date }
        weekday: { type: integer }
        lessons:
          type: array
          items: { $ref: "#/components/schemas/Lesson" }
        tasks:
          type: array
          items: { $ref: "#/components/schemas/Task" }
        tests:
          type: array
          items: { $ref: "#/components/schemas/Test" }
        debts:
          type: array
          items: { $ref: "#/components/schemas/Debt" }
        loadScore: { type: integer }
      required: [date, weekday, lessons, tasks, tests, debts, loadScore]